Fix spectrum bin level so it doesn't vary with bin width
The bin power seems to vary based on bin width. I'm testing with an
arb set to 2.5 MHz at ranges from -50 to -10 dBm, and I see the
baseband and spectrum power match only when bin width is 40 Hz. Wider
bins show less power. Removing the scale by bin count factor seems to
fix this. Or it's closer--the baseband and powers readings now match.