Add RX888 config option to skip USB reset
I can't get any alternate FX3 firmware to work if the USB reset is
present. Franco's rx888_stream program does work with alternate FX3
firmware and it does _not_ do the USB reset. Weird.

I've added a "hack_no_usb_reset" config option with the default to perform
the USB reset. Not sure if this is the correct approach...comments?